Update RedditUpvotesNLinks.user.js
Edited code to wait two seconds so that upvotes are accounted for. For some reason, if the function ran before this, the upvotes looked to not register in the newly opened tabs, possibly because the script is going so fast or because of some other thing about javascript that I haven't tackled yet.